pi/requirements.txt
awsiotsdk>=1.22.0
ruuvitag-sensor>=2.3.0
